# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive fitness platform API for gym members and trainers",
        "title": "FitTrack+ API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "FitTrack+ Team",
            "url": "http://www.fittrackplus.com/support",
            "email": "support@fittrackplus.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/trainer/workouts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all workouts created by the current trainer with optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workouts"
                ],
                "summary": "Get workouts created by trainer (Trainer only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by specific member",
                        "name": "member_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (active, inactive)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/workout.WorkoutResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - Trainer only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/workouts": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a personalized workout session for an assigned member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workouts"
                ],
                "summary": "Create a new workout for a member (Trainer only)",
                "parameters": [
                    {
                        "description": "Workout details",
                        "name": "workout",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workout.CreateWorkoutRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/workout.WorkoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - Trainer only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/workouts/my-workouts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all workouts assigned to the current member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workouts"
                ],
                "summary": "Get workouts assigned to member (Member only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (active, inactive)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/workout.WorkoutResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - Member only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/workouts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a specific workout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workouts"
                ],
                "summary": "Get specific workout details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workout.WorkoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - Access denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update workout details and exercises",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workouts"
                ],
                "summary": "Update an existing workout (Trainer only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated workout details",
                        "name": "workout",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workout.CreateWorkoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workout.WorkoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - Trainer only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "workout.CreateExerciseRequest": {
            "type": "object",
            "required": [
                "exercise_id",
                "order",
                "reps",
                "rest_time",
                "sets"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "exercise_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "order": {
                    "type": "integer",
                    "minimum": 1
                },
                "reps": {
                    "type": "integer",
                    "minimum": 1
                },
                "rest_time": {
                    "type": "integer",
                    "minimum": 0
                },
                "sets": {
                    "type": "integer",
                    "minimum": 1
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "workout.CreateWorkoutRequest": {
            "type": "object",
            "required": [
                "difficulty",
                "estimated_duration",
                "exercises",
                "member_id",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "string"
                },
                "estimated_duration": {
                    "type": "integer",
                    "minimum": 1
                },
                "exercises": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/workout.CreateExerciseRequest"
                    }
                },
                "member_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "plan_id": {
                    "type": "integer"
                }
            }
        },
        "workout.ExerciseInfo": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "string"
                },
                "equipment": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "instructions": {
                    "type": "string"
                },
                "muscle_group": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tips": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
        "workout.PlanInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "goal_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "plan_type": {
                    "type": "string"
                }
            }
        },
        "workout.UserInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "workout.WorkoutExerciseResponse": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "exercise": {
                    "description": "Exercise details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/workout.ExerciseInfo"
                        }
                    ]
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "reps": {
                    "type": "integer"
                },
                "rest_time": {
                    "type": "integer"
                },
                "sets": {
                    "type": "integer"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "workout.WorkoutResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "string"
                },
                "estimated_duration": {
                    "type": "integer"
                },
                "exercises": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workout.WorkoutExerciseResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "member": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/workout.UserInfo"
                        }
                    ]
                },
                "member_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "plan": {
                    "$ref": "#/definitions/workout.PlanInfo"
                },
                "plan_id": {
                    "type": "integer"
                },
                "trainer": {
                    "$ref": "#/definitions/workout.UserInfo"
                },
                "trainer_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}